refactored extension.ts to make a queryRunner class
